manage errors exclusively through the custom dedicated handler class
